perf(#378): 로딩 최적화 Phase2 — 느린 mount 콜 afterPaint 지연
프로덕션 실측: KRX-ETF(ke) 12s+500·시장투자자(m) 6s+500이 mount 발사 = 체감지연 주범.
- useAfterIdle.js 공유 훅 신설(requestIdleCallback 후 enabled, 캐시 즉시·네트워크만 지연)
- useKrxEtf/MarketInvestorSection/useFearGreed: enabled afterIdle로 임계경로에서 제거
- useNewsQuery: 로컬 useAfterIdle → 공유 import
- ⚠️ useIndices는 헤더 임계+스냅샷 미포함이라 유지(critique 준수)
build·test 256 PASS.
